Fix data for outside of hour range
It can happen that a particular hour does not have any data. Then
loading the metrics will return data for a much later time. This was
causing many hour blocks with empty or repeated data.

Ignore this data for now. This is still very wasteful as the returned
data is loaded more than once.