Merge utilization and saturation graphs
This halves the number of SVGs, gets rid of the subgrids, and simplifies
the component structure.

Keep implementing the rulers through CSS, but as the ruler now goes
through the middle, these need some more tricks.

Draw the dotted line everywhere (with a low z-index), as that's more
efficient -- it just needs to happen once for the entire hour, instead
of for each minute.