Dynamically scale unbounded values
Keep track of the global maximum of unbounded values: load, disk IO and
network IO. Dynamically set the scale accordingly. This is a trade-off
between having to scale too often when new data arrives, and scaling too
aggressively which makes the graphs useless.

I started with rounding up to the next biggest power of 10:

    x => Math.pow(10, Math.ceil(Math.log10(x)))

but this is too aggressive -- in the worst case, it only uses 10%+ε of
the available range.

So let's start with rounding up to the next "full" number, i.e. which
has a single digit followed by all zeros. Even in the worst case this
still uses 90%+ε of the range. It is very prone to rescaling, though, so
if that is too annoying or a performance problem, we need to introduce
an additional scaling factor (like 1.5) there.