Drop unknown data points
Don't pre-fill the entire data hour with `null`s, as that causes the
minute slices to always have the full minute length. That in turn causes
the polygons to always be drawn completely, and thus going to 0 instead
of ending in the middle of the minute when the data ends.

There is no need to pre-initialize the array length. JavaScript is
perfectly happy about setting a previously unset array index, and fills
up the gaps (if any) with "empty" elements.