Redesign Disks metrics
Move the total size into a tooltip to conserve space.

If cockpit-storaged is available, turn the mount point usage info into
links to the Storage page. There is no specific subpage for each mount,
the main page has a "Filesystems" table with all of them.

Wrap the entire <Progress> as link, so that the tooltip works properly
both with mouse hovering and with keyboard focusing, and the tooltip
points to the progress bar instead of just the text label.

Fixes #43
Closes #69