Drop total size from RAM/Swap
This makes the card consistent with Disks and also avoids overly long
labels. The total size isn't that important for debugging, and available
size plus the progress bar should give enough intuition about the total
size already.

Make the Progress bars focusable, so that the tooltip can be shown with
keyboard navigation.